+++ docs/changelog.md
@@ -23,6 +23,7 @@ MML changes:
  * Fixed slurred note tracking after a subroutine call.  
    A `{}` portamento after an MML subroutine that ends with a slurred note now has the same
    behaviour as a `{}` portamento after a loop that ends with a slurred note.
+ * Fixed slurred note tracking at the start of a loop and after a `:` skip last loop command.
 
 
 Version 0.0.10